修订Glitch系列算法
修订Glitch系列算法

+++ Assets/X-PostProcessing/Effects/GlitchWaveJitter/Shader/GlitchWaveJitter.shader
@@ -35,7 +35,7 @@ Shader "Hidden/X-PostProcessing/Glitch/WaveJitter"
 			strength = 0.5 + 0.5 *cos(_Time.y * _Frequency);
 		#endif
 		
-		// --------------------------------Prepare UV----------------------------
+		// Prepare UV
 		float uv_y = i.texcoord.y * _Resolution.y;
 		float noise_wave_1 = snoise(float2(uv_y * 0.01, _Time.y * _Speed * 20)) * (strength * _Amount * 32.0);
 		float noise_wave_2 = snoise(float2(uv_y * 0.02, _Time.y * _Speed * 10)) * (strength * _Amount * 4.0);
@@ -44,11 +44,11 @@ Shader "Hidden/X-PostProcessing/Glitch/WaveJitter"
 
 		float rgbSplit_uv_x = (_RGBSplit * 50 + (20.0 * strength + 1.0)) * noise_wave_x / _Resolution.x;
 
-		// -------------------------------Sample RGB Color------------------------------
-		float4 colorG = SAMPLE_TEXTURE2D(_MainTex, sampler_MainTex, float2(uv_x, i.texcoord.y));
-		float4 colorRB = SAMPLE_TEXTURE2D(_MainTex, sampler_MainTex, float2(uv_x + rgbSplit_uv_x, i.texcoord.y));
+		// Sample RGB Color-
+		half4 colorG = SAMPLE_TEXTURE2D(_MainTex, sampler_MainTex, float2(uv_x, i.texcoord.y));
+		half4 colorRB = SAMPLE_TEXTURE2D(_MainTex, sampler_MainTex, float2(uv_x + rgbSplit_uv_x, i.texcoord.y));
 		
-		return  float4(colorRB.r, colorG.g, colorRB.b, colorRB.a + colorG.a);
+		return  half4(colorRB.r, colorG.g, colorRB.b, colorRB.a + colorG.a);
 	}
 
 	float4 Frag_Vertical(VaryingsDefault i) : SV_Target
@@ -60,7 +60,7 @@ Shader "Hidden/X-PostProcessing/Glitch/WaveJitter"
 			strength = 0.5 + 0.5 * cos(_Time.y * _Frequency);
 		#endif
 
-		// --------------------------------Prepare UV--------------------------------
+		// Prepare UV
 		float uv_x = i.texcoord.x * _Resolution.x;
 		float noise_wave_1 = snoise(float2(uv_x * 0.01, _Time.y * _Speed * 20)) * (strength * _Amount * 32.0);
 		float noise_wave_2 = snoise(float2(uv_x * 0.02, _Time.y * _Speed * 10)) * (strength * _Amount * 4.0);
@@ -69,11 +69,11 @@ Shader "Hidden/X-PostProcessing/Glitch/WaveJitter"
 
 		float rgbSplit_uv_y = (_RGBSplit * 50 + (20.0 * strength + 1.0)) * noise_wave_y / _Resolution.y;
 
-		// -------------------------------Sample RGB Color------------------------------
-		float4 colorG = SAMPLE_TEXTURE2D(_MainTex, sampler_MainTex, float2(i.texcoord.x, uv_y));
-		float4 colorRB = SAMPLE_TEXTURE2D(_MainTex, sampler_MainTex, float2(i.texcoord.x, uv_y + rgbSplit_uv_y));
+		// Sample RGB Color
+		half4 colorG = SAMPLE_TEXTURE2D(_MainTex, sampler_MainTex, float2(i.texcoord.x, uv_y));
+		half4 colorRB = SAMPLE_TEXTURE2D(_MainTex, sampler_MainTex, float2(i.texcoord.x, uv_y + rgbSplit_uv_y));
 
-		return float4(colorRB.r, colorG.g, colorRB.b, colorRB.a + colorG.a);
+		return half4(colorRB.r, colorG.g, colorRB.b, colorRB.a + colorG.a);
 	}
 	
 	ENDHLSL
